proxy master/endpoint init is working
