+++ libs/block-mesh-common/Cargo.toml
@@ -8,3 +8,9 @@ authors.workspace = true
 
 [dependencies]
 tracing-subscriber = { workspace = true, features = ["env-filter", "time", "json"] }
+http-body-util = { workspace = true, optional = true }
+bytes = { workspace = true, optional = true }
+hyper = { workspace = true, features = ["full"], optional = true }
+
+[features]
+http = ["dep:http-body-util", "dep:bytes", "dep:hyper"]
