+++ libs/block-mesh-common/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod constants;
+pub mod http;
 pub mod tracing;
