+++ libs/test-proxy-master/src/clients_endpoint.rs
@@ -1,7 +1,7 @@
 use crate::proxy_pool::ProxyPool;
+use block_mesh_common::http::empty;
 use bytes::Bytes;
 use http_body_util::combinators::BoxBody;
-use http_body_util::{BodyExt, Empty};
 use hyper::service::service_fn;
 use hyper::{client, server, Method, Request, Response};
 use hyper_util::rt::TokioIo;
@@ -12,14 +12,13 @@ use tokio::net::TcpListener;
 pub async fn listen_for_clients_connecting(pool: ProxyPool, client_listener: TcpListener) {
     while let Ok((stream, addr)) = client_listener.accept().await {
         let pool = pool.clone();
-
         tokio::spawn(async move {
             if let Err(err) = server::conn::http1::Builder::new()
                 .preserve_header_case(true)
                 .title_case_headers(true)
                 .serve_connection(
                     TokioIo::new(stream),
-                    service_fn(move |req| handle_request(pool.clone(), req)),
+                    service_fn(move |req| handle_client_request(pool.clone(), req)),
                 )
                 .with_upgrades()
                 .await
@@ -30,14 +29,12 @@ pub async fn listen_for_clients_connecting(pool: ProxyPool, client_listener: Tcp
     }
 }
 
-#[tracing::instrument(name = "handle_request", ret, err)]
-pub async fn handle_request(
+#[tracing::instrument(name = "handle_client_request", ret, err)]
+async fn handle_client_request(
     pool: ProxyPool,
     mut req: Request<hyper::body::Incoming>,
 ) -> Result<Response<BoxBody<Bytes, hyper::Error>>, hyper::Error> {
     if Method::CONNECT == req.method() {
-        tracing::info!("CONNECT request from client");
-
         // Received an HTTP request like:
         // ```
         // CONNECT www.domain.com:443 HTTP/1.1
@@ -53,39 +50,27 @@ pub async fn handle_request(
         // `on_upgrade` future.
         tokio::spawn(async move {
             match hyper::upgrade::on(&mut req).await {
+                // TODO: can add headers here
                 Ok(upgraded) => {
                     let proxy = pool.get().await.unwrap();
                     let (mut send_request, conn) =
                         client::conn::http1::Builder::new().handshake(proxy).await?;
-
                     tokio::spawn(conn.with_upgrades());
                     let res = send_request.send_request(req).await?;
-
                     let stream = hyper::upgrade::on(res).await?;
-
-                    copy_bidirectional(&mut TokioIo::new(upgraded), &mut TokioIo::new(stream))
-                        .await
-                        .unwrap();
+                    let (from_client, from_server) =
+                        copy_bidirectional(&mut TokioIo::new(upgraded), &mut TokioIo::new(stream))
+                            .await
+                            .unwrap();
+                    tracing::info!(from_client, from_server);
                 }
-                Err(e) => eprintln!("upgrade error: {}", e),
+                Err(e) => tracing::error!("upgrade error = {}", e),
             }
             Ok::<(), hyper::Error>(())
         });
         Ok(Response::new(empty()))
     } else {
-        tracing::info!("NOT CONNECT request from client");
-
-        let proxy = pool.get().await.unwrap();
-        let (mut send_request, conn) = client::conn::http1::Builder::new().handshake(proxy).await?;
-
-        tokio::spawn(conn);
-
-        Ok(send_request.send_request(req).await?.map(|b| b.boxed()))
+        // TODO : Process request - can register client here
+        Ok(Response::new(empty()))
     }
-}
-
-fn empty() -> BoxBody<Bytes, hyper::Error> {
-    Empty::<Bytes>::new()
-        .map_err(|never| match never {})
-        .boxed()
 }
