+++ scripts/curl.sh
@@ -4,3 +4,8 @@ curl --proxy '192.155.93.107:3000' https://example.com
 
 # IPv6
 curl -6 --proxy '[2600:3c02::f03c:94ff:fef5:39d5]:3000' https://example.com
+
+
+curl http://httpbin.org/anything
+
+curl https://postman-echo.com/get
